@@ -1,7 +1,6 @@
 from setuptools import setup
 
-requires = ["six"]
-
+requires = []
 tests_require = ["mock", "pytest", "pytest-cov"]
 
 setup(
@@ -15,14 +14,12 @@ setup(
     packages=["findb"],
     zip_safe=False,
     install_requires=requires,
-    python_requires=">=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, <4",
+    python_requires=">=3.5, <4",
     tests_require=tests_require,
     classifiers=[
         "Natural Language :: English",
         "License :: OSI Approved :: MIT",
         "Programming Language :: Python",
-        "Programming Language :: Python :: 2",
-        "Programming Language :: Python :: 2.7",
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
